Automatically Committed for MakeGamesWithUs

+++ GameOfLife.spritebuilder/Grid.m
@@ -76,8 +76,8 @@ static const int GRID_COLUMNS = 10;
 - (Creature *)creatureForTouchPosition:(CGPoint)touchPosition
 {
     //get the row and column that was touched, return the Creature inside the corresponding cell
-    NSInteger row = touchPosition.y/_cellHeight;
-    NSInteger column = touchPosition.x/_cellWidth;
+    int row = touchPosition.y/_cellHeight;
+    int column = touchPosition.x/_cellWidth;
     return _gridArray[row][column];
 }
 @end
